backend/requirements.txt
# Core Frameworks
flask>=3.0.0
flask-cors>=4.0.0
waitress>=2.1.2             # Production server

# Computer Vision & AI
torch==2.1.0+cpu; sys_platform == 'win32'
torchvision==0.16.0+cpu; sys_platform == 'win32'
ultralytics>=8.0.0          # YOLOv8
opencv-python>=4.8.0        # Image processing
pillow>=10.0.0              # Image manipulation
scikit-learn>=1.3.0         # ML utilities
scipy>=1.10.0               # Scientific computing

# PDF Generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0        # Environment variables
pyyaml>=6.0                 # YAML config
tqdm>=4.65.0                # Progress bars
werkzeug>=3.0.0

# Optional AI
google-generativeai>=0.3.0  # Gemini API